make some existing logic and files easier to read
